developer report on package check process in shiny

developer report on package check process in shiny

browse_checks(con, pnet_zip)

Arguments

con

SQLiteConnection where tables of check results from rcmdcheck, etc., are stored

pnet_zip

character(1) path to a zip file holding pkgnet objects

Note

dbDisconnect is run if the stop button is pressed to end shiny session. If stopped via ctrl-C or error, the connection may remain and require closure. N.B. The source codes for this app are in inst/app1, to simplify contribution to shinyapps.io. Interactive execution proceeds by copying the server and ui files to temporary folder where shiny::runApp() is executed. For some packages for which tests do not succeed coverage statistics derived by pkgnet will be absent.

Examples

if (interactive()) {
 con = RSQLite::dbConnect(RSQLite::SQLite(), 
  system.file("sqlite/demo16.sqlite", package="BiocBuildTools"), flags=RSQLite::SQLITE_RO)
 print(RSQLite::dbListTables(con))
 #demo_pnet2 = readRDS(system.file("pkgnet/demo16_pnet.rds", package="BiocBuildTools"))
 demo_pnet2 = system.file("pkgnet/demo2.zip", package="BiocBuildTools")
 browse_checks(con, demo_pnet2)
}