a list of covr::package_coverage results, thinned for demonstration purposes, produced dec 2019

a list of covr::package_coverage results, thinned for demonstration purposes, produced dec 2019

cov_tab_demo

Format

list

Examples

cov_tab_demo[1:2]

#> $BiocFileCache
#> $BiocFileCache$overall
#> [1] 76.58582
#> 
#> $BiocFileCache$specific
#>        R/makeCachedActiveBinding.R                  R/sql_migration.R 
#>                            0.00000                           20.45455 
#>                           R/httr.R            R/BiocFileCache-class.R 
#>                           73.01587                           80.40816 
#>                      R/utilities.R R/makeBiocFileCacheFromDataFrame.R 
#>                           86.95652                           88.57143 
#>                            R/sql.R 
#>                           89.47368 
#> 
#> 
#> $bsseq
#> $bsseq$overall
#> [1] 34.12141
#> 
#> $bsseq$specific
#>       R/BSmooth.fstat.R       R/BSmooth.tstat.R     R/BSseqStat_class.R 
#>                 0.00000                 0.00000                 0.00000 
#>    R/BSseqTstat_class.R           R/dmrFinder.R            R/findLoci.R 
#>                 0.00000                 0.00000                 0.00000 
#>              R/fisher.R            R/getStats.R           R/gof_stats.R 
#>                 0.00000                 0.00000                 0.00000 
#>          R/hasGRanges.R            R/plotting.R        R/read.bsmooth.R 
#>                 0.00000                 0.00000                 0.00000 
#>     R/FWIRanges-class.R         R/BSseq_utils.R  R/DelayedArray_utils.R 
#>                14.28571                20.61856                29.31034 
#>        R/permutations.R               R/utils.R         R/BSseq-class.R 
#>                47.42268                56.00000                57.77778 
#>       R/collapseBSseq.R     R/FWGRanges-class.R          R/hdf5_utils.R 
#>                58.02469                69.00585                69.23077 
#>        R/read.bismark.R             R/BSmooth.R             R/combine.R 
#>                71.96262                82.35294                87.78626 
#> src/check_M_and_Cov.cpp            src/init.cpp           src/utils.cpp 
#>                90.47619               100.00000               100.00000 
#> 
#>