use pkgnet to analyze a source folder (including covr statistics)

use pkgnet to analyze a source folder (including covr statistics)

get_pnet(sourcedir, serialize = TRUE, overwrite = TRUE)

Arguments

sourcedir

character(1) should contain an R package

serialize

logical(1) defaults to TRUE, See note.

overwrite

logical(1) defaults to TRUE; if FALSE will fail if [basename(sourcedir)]_pnet.rds exists

Value

See Note.

Note

By default, serializes an S3 instance of BBT_pkgnet, a list with components dep for dependency report, and fun for function interrelationships, saved as RDS to [basename(sourcedir)]_pnet.rds. If serialize is FALSE, simply returns the object. This is intended to be used on a large number of packages and thus a checkpointing approach is the default.

Examples

td = tempdir()
cur = getwd()
setwd(td)
pdy = system.file("demo_srcs/parody", package="BiocBuildTools")
get_pnet(pdy)
#> Loading required package: pkgnet
#> INFO [2022-01-14 00:23:25] Creating graph model for network...
#> INFO [2022-01-14 00:23:25] Constructing dependency network for parody
#> INFO [2022-01-14 00:23:25] ...done constructing dependency network.
#> INFO [2022-01-14 00:23:25] ...graph model stored as pkg_graph.
#> INFO [2022-01-14 00:23:25] Calculating outDegree...
#> INFO [2022-01-14 00:23:25] Constructing igraph object...
#> INFO [2022-01-14 00:23:25] ...done constructing igraph object.
#> INFO [2022-01-14 00:23:25] Calculating inDegree...
#> INFO [2022-01-14 00:23:25] Calculating numRecursiveDeps...
#> INFO [2022-01-14 00:23:25] Calculating numRecursiveRevDeps...
#> INFO [2022-01-14 00:23:25] Calculating betweenness...
#> INFO [2022-01-14 00:23:25] Calculating pageRank...
#> INFO [2022-01-14 00:23:25] Calculating graphOutDegree
#> INFO [2022-01-14 00:23:25] Calculating graphInDegree
#> INFO [2022-01-14 00:23:25] Calculating graphBetweenness
#> INFO [2022-01-14 00:23:25] Calculating test coverage for parody...
#> INFO [2022-01-14 00:23:28] Updating cached nodes data.table with metadata...
#> INFO [2022-01-14 00:23:28] Extracting functions from parody as graph nodes...
#> Warning: there is no package called ‘parody’
#> Error in as.environment(pos) : 
#>   no item called "package:parody" on the search list
file.exists("parody_pnet.rds")
#> [1] TRUE
class(readRDS("parody_pnet.rds"))
#> [1] "BBT_pkgnet" "list"      
setwd(cur)